requirements.txt
# MIRA - Validated Third-Party Requirements
# Generated by analyzing actual usage across the codebase
# Each package verified to be actively used in non-deprecated files

# == PYTORCH CPU-ONLY INSTALLATION ==
--extra-index-url https://download.pytorch.org/whl/cpu

# == WEB FRAMEWORK & SERVER ==
fastapi           # Core web framework - heavily used across API endpoints
starlette         # FastAPI dependency - used in middleware
hypercorn         # ASGI server with HTTP/2 support - required for all environments

# == DATABASE & DATA PERSISTENCE ==
psycopg2-binary   # PostgreSQL driver (binary distribution) - postgres_client.py, database utils
pgvector          # PostgreSQL vector extension - postgres_client.py

# == CACHING & KEY-VALUE STORE ==
valkey            # Redis-compatible cache - valkey_client.py, many services

# == AUTHENTICATION & SECURITY ==
PyJWT             # JWT token handling - auth service, tests
cryptography      # Encryption utilities - credential_security.py, userdata_manager.py

# == AI & MACHINE LEARNING ==
anthropic        # Anthropic Claude API SDK - primary LLM provider with prompt caching
openai           # OpenAI API client - tools, embeddings (fallback for some tools)
numpy            # Numerical computing - embeddings, vector operations, ML utilities
torch            # PyTorch (CPU-only via index above) - sentence_transformers.py, bge_embeddings.py
transformers     # Hugging Face transformers - embeddings, BGE models
sentence-transformers  # Sentence embeddings - all-MiniLM, BGE reranker models
optimum[onnxruntime]  # ONNX model optimization - sentence_transformers.py, BGE embeddings
onnx             # ONNX runtime - sentence_transformers.py, bge_embeddings.py
onnxruntime      # ONNX inference engine - bge_embeddings.py
onnxconverter_common  # ONNX conversion utilities - bge_embeddings.py
onnxscript           # ONNX model scripting - bge_reranker.py torch.onnx.export
tqdm             # Progress bars - bge_embeddings.py model loading
spacy>=3.7.0     # NER and NLP - entity extraction for memory linking
# Note: Install en_core_web_lg with: python -m spacy download en_core_web_lg

# == SCHEDULING & BACKGROUND TASKS ==
apscheduler      # Task scheduling - memory_extraction_service.py, auth_service.py, scheduler_service.py

# == HTTP CLIENT & NETWORKING ==
httpx[http2]     # Modern HTTP client with HTTP/2 support - tools, LLM provider, embeddings, weather/web tools
aiohttp          # Async HTTP - kasa_tool.py for smart home control
mcp              # Model Context Protocol SDK - mcp_client.py for MCP server connections

# == TOOL-SPECIFIC LIBRARIES ==
googlemaps       # Google Maps API - maps_tool.py
kasa             # TP-Link Kasa smart devices - kasa_tool.py
kagiapi          # Kagi search API - webaccess tools
hvac             # HashiCorp Vault client - vault_client.py

# == IMAGE PROCESSING ==
Pillow>=10.0.0   # Image compression - two-tier compression for multimodal messages

# == UTILITIES & HELPERS ==
pydantic         # Data validation - widely used across API models and tools
python-dateutil  # Date parsing - reminder_tool.py, timezone_utils.py
pytz             # Timezone handling - timezone_utils.py
colorama         # Terminal colors - colored_logging.py (already in requirements.txt)
rich             # Rich text and TUI components - talkto_mira.py CLI interface
rapidfuzz        # Fuzzy string matching - memory deduplication, entity normalization
json-repair      # JSON repair for malformed LLM responses - extraction.py
psutil           # Process management - playwright_service.py for forced browser shutdown

# == TESTING DEPENDENCIES ==
pytest           # Test framework - extensive test suite
pytest-asyncio   # Async test support - async tests

# == OPTIONAL DEPENDENCIES ==
# These packages enhance functionality but are not required for core operation
# Install with: pip install python-docx openpyxl playwright

# Document processing - enables richer DOCX/XLSX text extraction
# Falls back to stdlib XML parsing if not installed
python-docx      # (optional) DOCX text extraction - document_processing.py
openpyxl         # (optional) XLSX text extraction - document_processing.py

# Web scraping - enables JavaScript-rendered page capture
playwright       # (optional) Headless browser - web_tool.py

# == PACKAGES TO REVIEW/CONSIDER REMOVING ==
# These packages have limited usage and may be removable:
# - pytz (could potentially use zoneinfo from Python 3.9+ instead)
